Pythagoras tree: Merge two widgets into single widget
